+++ Backend/lin/entrypoint.sh
@@ -1,3 +1,3 @@
-python sources/manage.py makemigrations
-python sources/manage.py migrate
-python sources/manage.py runserver 0.0.0.0:8088
+python ../sources/manage.py makemigrations core app
+python ../sources/manage.py migrate
+python ../sources/manage.py runserver 0.0.0.0:8088