+++ Backend/sources/apps/app/ws_routing.py
@@ -4,3 +4,5 @@ from . import ws_consumers
 websocket_urlpatterns = [
     re_path(r'ws/chat/(?P<room_uuid>[a-f0-9]{8}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{12})/$', ws_consumers.ChatConsumer.as_asgi()),
 ]
+
+