+++ Backend/sources/project/asgi.py
@@ -13,12 +13,13 @@ from django.core.asgi import get_asgi_application
 from channels.routing import ProtocolTypeRouter, URLRouter
 from channels.auth import AuthMiddlewareStack
 from apps.app import ws_routing
+from apps.app.ws_consumers import TokenAuthMiddleware
 
 os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'project.settings')
 
 application = ProtocolTypeRouter({
     'http': get_asgi_application(),
-    'websocket':AuthMiddlewareStack(
+    'websocket':TokenAuthMiddleware(
         URLRouter(
             ws_routing.websocket_urlpatterns
         )